Create new folder structure for models
