+++ apps/recipe_api/admin.py
@@ -1,10 +1,30 @@
 from django.contrib import admin
-from apps.recipe_api.models import (
-    Category, Recipe, Ingredient, Instruction
-)
+from django.contrib.admin.models import LogEntry
+from apps.recipe_api.models.category import Category
+from apps.recipe_api.models.recipe import Recipe
+from apps.recipe_api.models.ingredient import Ingredient
+from apps.recipe_api.models.instruction import Instruction
 
 # Register your models here.
-admin.site.register(Category)
+@admin.register(LogEntry)
+class LogEntryAdmin(admin.ModelAdmin):
+    def has_change_permission(self, request, obj = None):
+        return False
+    
+    def has_delete_permission(self, request, obj = None):
+        return False
+        
+    def has_add_permission(self, request):
+        return None
+    
+class CategoryCustomAdmin(admin.ModelAdmin):
+    search_fields = ['name']
+    show_full_result_count = True
+    list_filter = ['name']
+    list_display = ['name']
+
+admin.site.register(Category, CategoryCustomAdmin)
+
 admin.site.register(Recipe)
 admin.site.register(Ingredient)
 admin.site.register(Instruction)
