+++ apps/recipe_api/models/__init__.py
@@ -1,0 +1,4 @@
+from apps.recipe_api.models import category
+from apps.recipe_api.models import recipe
+from apps.recipe_api.models import ingredient
+from apps.recipe_api.models import instruction
