+++ apps/recipe_api/models/category.py
@@ -1,0 +1,15 @@
+from django.db import models
+from apps.recipe_api.timestamp import BaseTimeStampModel
+
+# Create your models here.
+class Category(BaseTimeStampModel):
+    """
+    Represents a category for recipes (e.g., 'Dessert', 'Main Course', 'Breakfast').
+    """
+    name = models.CharField(max_length=100, unique=True)
+
+    class Meta:
+        verbose_name_plural = "Categories"
+
+    def __str__(self):
+        return self.name
