+++ apps/recipe_api/models/instruction.py
@@ -1,0 +1,26 @@
+from django.db import models
+from apps.recipe_api.models.recipe import Recipe
+from apps.recipe_api.timestamp import BaseTimeStampModel
+from django.core.validators import MinValueValidator
+
+class Instruction(BaseTimeStampModel):
+    """
+    Represents step-by-step instructions for a recipe.
+    """
+    recipe = models.ForeignKey(
+        Recipe, 
+        on_delete=models.CASCADE,
+        related_name='instructions'
+    )
+    step_number = models.PositiveIntegerField(
+        validators=[MinValueValidator(1)],
+        help_text='Order of the instruction step'
+    )
+    description = models.TextField()
+
+    class Meta:
+        ordering = ['step_number']
+        unique_together = ('recipe', 'step_number')
+
+    def __str__(self):
+        return f"Step {self.step_number} for {self.recipe.title}"
